requirements.txt
# LEED Certification Automation Platform Requirements
# Based on the research paper's technology stack

# Core dependencies
pdfplumber>=0.9.0
dataclasses>=0.6
typing-extensions>=4.0.0

# Energy modeling (EnergyPlus integration)
eppy>=0.5.56

# LLM and RAG components
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
rank-bm25>=0.2.0

# GIS and location analysis
geopandas>=0.13.0
pandas>=2.0.0
shapely>=2.0.0

# Optional OCR components
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=9.5.0

# HTTP requests for APIs
requests>=2.31.0

# Logging and utilities
python-dateutil>=2.8.0

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# Graphviz DOT parsing
graphviz>=0.21
pydot>=4.0.1

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 